# api/docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Artificialic Budget Platform API",
    "description": "API REST provista por nosotros. Consumida por Artificialic (ingest) y dashboard interno.",
    "version": "1.0.0"
  },
  "paths": {
...
    "/v1/budgets": {
      "post": {
        "tags": [
          "ingest"
        ],
        "summary": "Ingest de presupuesto (desde chatbot)",
        "description": "Endpoint p\u00fablico consumido por los chatbots de Artificialic. Requiere API Key con scope `budgets:write`.",
        "operationId": "create_budget_v1_budgets_post",
        "parameters": [
          {
            "name": "X-API-Key",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Api-Key"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BudgetCreate"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BudgetOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "dashboard"
        ],
        "summary": "Listar presupuestos (paginado + filtros)",
        "operationId": "list_budgets_v1_budgets_get",
        "parameters": [
          {
            "name": "status",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Status"
            }
          },
          {
            "name": "priority",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Priority"
            }
          },
          {
            "name": "source",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Source"
            }
          },
          {
            "name": "assigned_to",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Assigned To"
            }
          },
          {
            "name": "search",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "minLength": 1
                },
                {
                  "type": "null"
                }
              ],
              "title": "Search"
            }
          },
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 1,
              "default": 1,
              "title": "Page"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100,
              "minimum": 1,
              "default": 20,
              "title": "Limit"
            }
          },
          {
            "name": "sort_by",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "default": "created_at",
              "title": "Sort By"
            }
          },
          {
            "name": "sort_order",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "pattern": "^(asc|desc)$",
              "default": "desc",
              "title": "Sort Order"
            }
          },
          {
            "name": "X-API-Key",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Api-Key"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BudgetListOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/budgets/{budget_id}": {
      "get": {
        "tags": [
          "dashboard"
        ],
        "summary": "Detalle de presupuesto (incluye notas y history)",
        "operationId": "get_budget_v1_budgets__budget_id__get",
        "parameters": [
          {
            "name": "budget_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Budget Id"
            }
          },
          {
            "name": "X-API-Key",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Api-Key"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BudgetOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "patch": {
        "tags": [
          "dashboard"
        ],
        "summary": "Actualizar estado, prioridad o asignado",
        "description": "Requiere scope `budgets:update`. Valida la transici\u00f3n de estado contra la m\u00e1quina de estados del proyecto; si la transici\u00f3n va a `cancelado` o `rechazado`, `reason` es obligatorio.",
        "operationId": "patch_budget_v1_budgets__budget_id__patch",
        "parameters": [
          {
            "name": "budget_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Budget Id"
            }
          },
          {
            "name": "X-API-Key",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Api-Key"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BudgetPatch"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BudgetOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/budgets/{budget_id}/notes": {
      "post": {
        "tags": [
          "dashboard"
        ],
        "summary": "Agregar nota interna al presupuesto",
        "operationId": "add_note_v1_budgets__budget_id__notes_post",
        "parameters": [
          {
            "name": "budget_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Budget Id"
            }
          },
          {
            "name": "X-API-Key",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Api-Key"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/NoteCreate"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/NoteOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "BudgetCreate": {
        "properties": {
          "external_id": {
            "type": "string",
            "maxLength": 128,
            "minLength": 1,
            "title": "External Id",
            "description": "ID en el sistema de Artificialic (idempotencia)"
          },
          "client_name": {
            "type": "string",
            "maxLength": 255,
            "minLength": 1,
            "title": "Client Name"
          },
          "client_email": {
            "type": "string",
            "format": "email",
            "title": "Client Email"
          },
          "client_phone": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 64
              },
              {
                "type": "null"
              }
            ],
            "title": "Client Phone"
          },
          "client_company": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 255
              },
              {
                "type": "null"
              }
            ],
            "title": "Client Company"
          },
          "service_type": {
            "type": "string",
            "maxLength": 128,
            "minLength": 1,
            "title": "Service Type"
          },
          "description": {
            "type": "string",
            "minLength": 1,
            "title": "Description"
          },
          "estimated_amount": {
            "anyOf": [
              {
                "type": "number",
                "exclusiveMinimum": 0.0
              },
              {
                "type": "string",
                "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d{0,2}0*$"
              }
            ],
            "title": "Estimated Amount"
          },
          "currency": {
            "type": "string",
            "maxLength": 3,
            "minLength": 3,
            "title": "Currency",
            "default": "USD"
          },
          "priority": {
            "type": "string",
            "enum": [
              "low",
              "medium",
              "high",
              "urgent"
            ],
            "title": "Priority",
            "default": "medium"
          },
          "source": {
            "type": "string",
            "enum": [
              "whatsapp",
              "web",
              "messenger",
              "email"
            ],
            "title": "Source"
          },
          "tags": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "maxItems": 32,
            "title": "Tags"
          },
          "due_date": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Due Date"
          }
        },
        "type": "object",
        "required": [
          "external_id",
          "client_name",
          "client_email",
          "service_type",
          "description",
          "estimated_amount",
          "source"
        ],
        "title": "BudgetCreate",
        "description": "Payload que los chatbots de Artificialic env\u00edan a POST /v1/budgets."
      },
      "BudgetListMeta": {
        "properties": {
          "total": {
            "type": "integer",
            "title": "Total"
          },
          "page": {
            "type": "integer",
            "title": "Page"
          },
          "limit": {
            "type": "integer",
            "title": "Limit"
          },
          "total_pages": {
            "type": "integer",
            "title": "Total Pages"
          }
        },
        "type": "object",
        "required": [
          "total",
          "page",
          "limit",
          "total_pages"
        ],
        "title": "BudgetListMeta"
      },
      "BudgetListOut": {
        "properties": {
          "data": {
            "items": {
              "$ref": "#/components/schemas/BudgetOut"
            },
            "type": "array",
            "title": "Data"
          },
          "meta": {
            "$ref": "#/components/schemas/BudgetListMeta"
          }
        },
        "type": "object",
        "required": [
          "data",
          "meta"
        ],
        "title": "BudgetListOut"
      },
      "BudgetOut": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "external_id": {
            "type": "string",
            "title": "External Id"
          },
          "client_name": {
            "type": "string",
            "title": "Client Name"
          },
          "client_email": {
            "type": "string",
            "format": "email",
            "title": "Client Email"
          },
          "client_phone": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Client Phone"
          },
          "client_company": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Client Company"
          },
          "service_type": {
            "type": "string",
            "title": "Service Type"
          },
          "description": {
            "type": "string",
            "title": "Description"
          },
          "estimated_amount": {
            "type": "string",
            "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$",
            "title": "Estimated Amount"
          },
          "currency": {
            "type": "string",
            "title": "Currency"
          },
          "status": {
            "type": "string",
            "enum": [
              "recibido",
              "en_revision",
              "cotizado",
              "negociando",
              "aprobado",
              "en_proceso",
              "completado",
              "cancelado",
              "rechazado"
            ],
            "title": "Status"
          },
          "priority": {
            "type": "string",
            "enum": [
              "low",
              "medium",
              "high",
              "urgent"
            ],
            "title": "Priority"
          },
          "source": {
            "type": "string",
            "enum": [
              "whatsapp",
              "web",
              "messenger",
              "email"
            ],
            "title": "Source"
          },
          "assigned_to": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Assigned To"
          },
          "tags": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Tags"
          },
          "due_date": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Due Date"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Updated At"
          },
          "created_by_key_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Created By Key Id"
          },
          "notes": {
            "items": {
              "$ref": "#/components/schemas/NoteOut"
            },
            "type": "array",
            "title": "Notes",
            "default": []
          },
          "status_changes": {
            "items": {
              "$ref": "#/components/schemas/StatusChangeOut"
            },
            "type": "array",
            "title": "Status Changes",
            "default": []
          }
        },
        "type": "object",
        "required": [
          "id",
          "external_id",
          "client_name",
          "client_email",
          "service_type",
          "description",
          "estimated_amount",
          "currency",
          "status",
          "priority",
          "source",
          "tags",
          "created_at",
          "updated_at"
        ],
        "title": "BudgetOut"
      },
      "BudgetPatch": {
        "properties": {
          "status": {
            "anyOf": [
              {
                "type": "string",
                "enum": [
                  "recibido",
                  "en_revision",
                  "cotizado",
                  "negociando",
                  "aprobado",
                  "en_proceso",
                  "completado",
                  "cancelado",
                  "rechazado"
                ]
              },
              {
                "type": "null"
              }
            ],
            "title": "Status"
          },
          "priority": {
            "anyOf": [
              {
                "type": "string",
                "enum": [
                  "low",
                  "medium",
                  "high",
                  "urgent"
                ]
              },
              {
                "type": "null"
              }
            ],
            "title": "Priority"
          },
          "assigned_to": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Assigned To"
          },
          "reason": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Reason",
            "description": "Obligatoria para status cancelado/rechazado"
          }
        },
        "type": "object",
        "title": "BudgetPatch"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "NoteCreate": {
        "properties": {
          "content": {
            "type": "string",
            "minLength": 1,
            "title": "Content"
          },
          "author": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 255
              },
              {
                "type": "null"
              }
            ],
            "title": "Author"
          }
        },
        "type": "object",
        "required": [
          "content"
        ],
        "title": "NoteCreate"
      },
      "NoteOut": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "author": {
            "type": "string",
            "title": "Author"
          },
          "content": {
            "type": "string",
            "title": "Content"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "author",
          "content",
          "created_at"
        ],
        "title": "NoteOut"
      },
      "StatusChangeOut": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "from_status": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "From Status"
          },
          "to_status": {
            "type": "string",
            "title": "To Status"
          },
          "changed_by": {
            "type": "string",
            "title": "Changed By"
          },
          "reason": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Reason"
          },
          "changed_at": {
            "type": "string",
            "format": "date-time",
            "title": "Changed At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "from_status",
          "to_status",
          "changed_by",
          "changed_at"
        ],
        "title": "StatusChangeOut"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}